Communication of wheel controller working, controller parameters itself need further testing

+++ include/motor.h
@@ -20,12 +20,12 @@ class Motor{
         uint8_t resolution;
         uint8_t duty_cycle = 0;
 
-        void setOutput(uint8_t dirA, uint8_t dirB, uint8_t duty_cycle);
+        void setOutput(uint8_t dirA, uint8_t dirB, float duty_cycle);
     public:
         Motor(motor_config_t conf, ledc_channel_t pwm_channel, uint freq, uint8_t resolution);
         bool init();
-        void rotateCW(uint8_t duty_cycle);
-        void rotateCCW(uint8_t duty_cycle);
+        void rotateCW(float duty_cycle);
+        void rotateCCW(float duty_cycle);
         void stop();
         uint8_t getDutyCycle();
         bool isSaturated();
